Change max_tries/max_simultaneous from u32 to i32 in API types and config (p-fencyuhlh)
Unify integer types so the API-facing types (AgentRecord, UpsertAgentRequest) and config constants match the domain Agent struct (i32), eliminating all as-i32/as-u32 boundary casts in routes/agents.rs, spawner.rs, and run_spawners.rs. No user-visible change. All 648 tests pass, clippy and fmt clean.

+++ metis-server/src/test/agents.rs
@@ -28,7 +28,7 @@ async fn test_state_with_agents(agent_names: &[&str]) -> TestStateHandles {
             name.to_string(),
             format!("/agents/{name}/prompt.md"),
             3,
-            DEFAULT_AGENT_MAX_SIMULTANEOUS as i32,
+            DEFAULT_AGENT_MAX_SIMULTANEOUS,
             false,
         );
         store.add_agent(agent).await.unwrap();
